proyecto_salon/salon_managment/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
python-dotenv==1.0.0
PyMySQL==1.1.0
marshmallow==3.20.1
flask-marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0
Werkzeug==3.0.1
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.6.0
cloudinary==1.41.0
